lambda/requirements.txt
fastapi>=0.104.1
mangum>=0.17.0
psycopg2-binary>=2.9.9
pydantic>=2.5.0

